Add '!honkey' game
When a non-bot viewer or broadcaster types in chat, the current time is
recorded. This compiles a mapping of active chatters and the time of
their most recent message. The 'honkey' game considers all chatters
active within the last 5 minutes and randomly selects one to be
interacted with by the caller.